pyproject.toml
[tool.poetry]
name = "poetry-FlexibleCompanyBot"
version = "0.1.0"
description = ""
authors = ["Francisco Mouriz Sierra"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "4.6.1"
fastapi = "^0.95.1"
uvicorn = "^0.21.0"
python-docx = "^0.8.11"
beautifulsoup4 = "^4.12.0"
tqdm = "^4.56.0"
spacy = "^3.5.1"
torch = "2.0.0"
pyserini = "0.20.0"
language-tool-python = "2.7.1"
lxml = "4.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
